[REFACTOR] actions in features and steps files
