+++ test/acceptance/behave/components/common_steps/initial_steps.py
@@ -22,17 +22,23 @@
 """
 __author__ = 'Iván Arias León (ivan dot ariasleon at telefonica dot com)'
 
-import behave
-from behave import step
+#import behave
+#from behave import step
 
-from components.common_steps.requests import *
+from components.common_steps.general_steps import *
 from iotqatools.fabric_utils import FabricSupport
 from iotqatools.mongo_utils import Mongo
 from iotqatools.cb_v2_utils import CB
 from iotqatools.helpers_utils import *
 
-
 from tools.properties_config import Properties  # methods in properties class
+
+# constants
+CONTEXT_BROKER_ENV = u'context_broker_env'
+MONGO_ENV = u'mongo_env'
+properties_class = Properties()
+props_cb = properties_class.read_properties()[CONTEXT_BROKER_ENV]       # context broker properties dict
+props_mongo = properties_class.read_properties()[MONGO_ENV]             # mongo properties dict
 
 behave.use_step_matcher("re")
 __logger__ = logging.getLogger("steps")
@@ -42,67 +48,70 @@ __logger__ = logging.getLogger("steps")
 def update_properties_file(context, properties_file, sudo_run):
     """
     update properties.py file from setting folder or jenkins console
-    :param context:
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     :param properties_file: file to get data to update properties.py
     """
-    global properties_class
-    properties_class = Properties()
+    __logger__.debug("Updating properties.json file...")
     properties_class.update_properties_json_file(properties_file, sudo_run)
+    __logger__.debug("...Updated properties.json file")
 
 
-@step(u'update contextBroker config file and restart service')
+@step(u'update contextBroker config file')
 def update_context_broker_config_file_and_restart_service(context):
     """
-    updating /etc/sysconfig/contextBroker file an restarting service
-    :param context:
+    updating /etc/sysconfig/contextBroker file
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     """
-    global properties_class, props_cb, props_mongo, my_fab, configuration
-    __logger__.debug(" >> updating /etc/sysconfig/contextBroker file")
-    props = properties_class.read_properties()  # properties dict
-    props_cb = props["context_broker_env"]  # context broker properties dict
     __logger__.debug(" Context Broker parameters:")
     for param in props_cb:
         __logger__.debug("   %s: %s" % (param, props_cb[param]))
-    props_mongo = props["mongo_env"]  # mongo properties dict
     __logger__.debug(" Mongo parameters:")
     for param in props_mongo:
         __logger__.debug("   %s: %s" % (param, props_mongo[param]))
-
-    my_fab = FabricSupport(host=props_cb["CB_HOST"], user=props_cb["CB_FABRIC_USER"],
+    context.my_fab = FabricSupport(host=props_cb["CB_HOST"], user=props_cb["CB_FABRIC_USER"],
                            password=props_cb["CB_FABRIC_PASS"], cert_file=props_cb["CB_FABRIC_CERT"],
                            retry=props_cb["CB_FABRIC_RETRY"], hide=True, sudo=props_cb["CB_FABRIC_SUDO"])
-    configuration = properties_class.read_configuration_json()
-    __logger__.debug("CB_RUNNING_MODE: %s" % configuration["CB_RUNNING_MODE"])
-    if configuration["CB_RUNNING_MODE"].upper() == "RPM":
-        properties_class.update_context_broker_file(my_fab)
-        __logger__.info(" >> updated /etc/sysconfig/contextBroker file")
-        __logger__.debug(" >> restarting contextBroker service")
-        my_fab.run("service contextBroker restart")
-        __logger__.info(" >> restarted contextBroker service")
+    context.configuration = properties_class.read_configuration_json()
+    __logger__.debug("CB_RUNNING_MODE: %s" % context.configuration["CB_RUNNING_MODE"])
+    if context.configuration["CB_RUNNING_MODE"].upper() == "RPM":
+        __logger__.debug("Updating /etc/sysconfig/contextBroker file...")
+        properties_class.update_context_broker_file(context.my_fab)
+        __logger__.info("...Updated /etc/sysconfig/contextBroker file")
+
+
+@step(u'start ContextBroker')
+def start_context_broker(context):
+    """
+    start ContextBroker
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
+    """
+    if context.configuration["CB_RUNNING_MODE"].upper() == "RPM":
+        __logger__.debug("Starting contextBroker service...")
+        context.my_fab.run("service contextBroker restart")
+        __logger__.info("...Started contextBroker service")
     else:
-        __logger__.debug(" >> restarting contextBroker per command line interface")
+        __logger__.debug("Starting contextBroker per command line interface...")
         props_cb["CB_EXTRA_OPS"] = props_cb["CB_EXTRA_OPS"].replace('"', "")
         # hint: the -harakiri option is used to kill contextBroker (must be compiled in DEBUG mode)
         __logger__.debug("contextBroker -port %s -logDir %s -pidpath %s -dbhost %s -db %s %s -harakiri" %
             (props_cb["CB_PORT"], props_cb["CB_LOG_FILE"], props_cb["CB_PID_FILE"], props_mongo["MONGO_HOST"],
              props_mongo["MONGO_DATABASE"], props_cb["CB_EXTRA_OPS"]))
-        resp = my_fab.run("contextBroker -port %s -logDir %s -pidpath %s -dbhost %s -db %s %s -harakiri" %
+        resp = context.my_fab.run("contextBroker -port %s -logDir %s -pidpath %s -dbhost %s -db %s %s -harakiri" %
             (props_cb["CB_PORT"], props_cb["CB_LOG_FILE"], props_cb["CB_PID_FILE"], props_mongo["MONGO_HOST"],
              props_mongo["MONGO_DATABASE"], props_cb["CB_EXTRA_OPS"]))
-        __logger__.debug("output: %s" % str(resp))
-        __logger__.info(" >> restarted contextBroker command line interface")
+        __logger__.debug("output: %s" % repr(resp))
+        __logger__.info("...Started contextBroker command line interface")
 
 
-@step(u'stop service')
-def stop_service(context):
+@step(u'stop ContextBroker')
+def stop_Context_broker(context):
     """
-    stop ContextBroker service
-    :param context:
+    stop ContextBroker
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     """
-    global my_fab, configuration
-    if configuration["CB_RUNNING_MODE"].upper() == "RPM":
+    if context.configuration["CB_RUNNING_MODE"].upper() == "RPM":
         __logger__.debug("Stopping contextBroker service...")
-        my_fab.run("service contextBroker stop")
+        context.my_fab.run("service contextBroker stop")
         __logger__.info("...Stopped contextBroker service")
     else:
         __logger__.debug("Stopping contextBroker per harakiri...")
@@ -115,10 +124,9 @@ def stop_service(context):
 def verify_context_broker_is_installed_successfully(context):
     """
     verify contextBroker is started successfully
-    :param context:
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     """
-    global props_cb
-    __logger__.debug(" >> verify if contextBroker is started successfully")
+    __logger__.debug("Verifying if contextBroker is started successfully...")
     cb = CB(protocol=props_cb["CB_PROTOCOL"], host=props_cb["CB_HOST"], port=props_cb["CB_PORT"])
     c = 0
     while (not cb.is_cb_started()) and (int(props_cb["CB_RETRIES"]) > c):
@@ -131,21 +139,20 @@ def verify_context_broker_is_installed_successfully(context):
         context.execute_steps(u'Given verify if version is the expected')
     else:
         __logger__.info("Current version is not verified...")
-    __logger__.info(" >> verified that contextBroker is started successfully")
+    __logger__.info("...Verified that contextBroker is started successfully")
 
 
 @step(u'verify mongo is installed successfully')
 def verify_mongo_is_installed_successfully(context):
     """
     verify contextBroker is installed successfully
-    :param context:
+    :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     """
-    global props_mongo
-    __logger__.debug(" >> verify if mongo is installed successfully")
-    m = Mongo(host=props_mongo["MONGO_HOST"], port=props_mongo["MONGO_PORT"], user=props_mongo["MONGO_USER"],
-              password=props_mongo["MONGO_PASS"], version=props_mongo["MONGO_VERSION"],
-              verify_version=props_mongo["MONGO_VERIFY_VERSION"])
-    m.connect()
-    m.eval_version()
-    m.disconnect()
-    __logger__.info(" >> verified that mongo is installed successfully")
+    __logger__.debug("Verifying if mongo is installed successfully...")
+    mongo = Mongo(host=props_mongo["MONGO_HOST"], port=props_mongo["MONGO_PORT"], user=props_mongo["MONGO_USER"],
+                  password=props_mongo["MONGO_PASS"], version=props_mongo["MONGO_VERSION"],
+                  verify_version=props_mongo["MONGO_VERIFY_VERSION"])
+    mongo.connect()
+    mongo.eval_version()
+    mongo.disconnect()
+    __logger__.info("...Verified that mongo is installed successfully")
