+++ test/acceptance/behave/components/ngsiv2/entities/update_append_attribute_by_id.feature
@@ -33,11 +33,18 @@ Feature: update or append an attribute by entity ID using NGSI v2. "POST" - /v2/
   I would like to update or append an attribute by entity ID
   So that I can manage and use them in my scripts
 
-  BackgroundFeature:
-  Setup: update properties test file from "epg_contextBroker.txt" and sudo local "false"
-  Setup: update contextBroker config file and restart service
-  Check: verify contextBroker is installed successfully
-  Check: verify mongo is installed successfully
+  Actions Before the Feature:
+     Setup: update properties test file from "epg_contextBroker.txt" and sudo local "false"
+     Setup: update contextBroker config file
+     Setup: start ContextBroker
+     Check: verify contextBroker is installed successfully
+     Check: verify mongo is installed successfully
+
+  Actions After each Scenario:
+     Setup: delete database in mongo
+
+  Actions After the Feature:
+     Setup: stop ContextBroker
 
   @happy_path
   Scenario:  update and append attributes by entity ID using NGSI v2
