+++ test/acceptance/behave/environment.py
@@ -27,10 +27,20 @@ import os
 from termcolor import colored
 
 from components.common_steps.initial_steps import *
-from components.common_steps.requests import *
+from components.common_steps.general_steps import *
 
 __logger__ = logging.getLogger("environment")
 
+
+# constants
+ACTIONS_BEFORE_FEATURE = u'actions before the feature'
+ACTIONS_AFTER_FEATURE = u'actions after the feature'
+ACTIONS_BEFORE_SCENARIO = u'actions before each scenario'
+ACTIONS_AFTER_SCENARIO = u'actions after each scenario'
+KEYWORDS = ["Setup", "Check"]
+GIVEN_PREFIX = u'Given'
+AND_PREFIX = u'And'
+SEPARATOR = u':'
 
 def __create_log_folder(name):
     """
@@ -43,6 +53,53 @@ def __create_log_folder(name):
             __logger__.info("log folder has been created with name: %s" % name)
     except Exception, e:
         assert False, "ERROR  - creating logs folder \n       - %s" % str(e)
+
+
+def get_steps_from_feature_description(label, feature):
+    """
+    return all steps defined in feature description associated to label
+    :param label: labels allowed:
+       - Actions Before the Feature
+       - Actions Before each Scenario
+       - Actions After each Scenario
+       - Actions After the Feature
+    :param feature: feature values
+    :return: list
+    Hint: must be ":" in the step prefix (used as separator). ex: "Setup: "
+    """
+    steps_list = []
+    label_exists = False
+    for description in feature.description:
+        if label_exists:
+            if description.split(SEPARATOR)[0] in KEYWORDS:
+                steps_list.append(description.replace(description.split(SEPARATOR)[0]+SEPARATOR, GIVEN_PREFIX))
+            else:
+                break
+        if description.lower().find(label) >= 0:
+            label_exists = True
+    return steps_list
+
+
+def replace_given_to_add(value):
+    """
+    replacing Given prefix to Add prefix
+    :param value: text to verify
+    :return: string
+    """
+    return value.replace(GIVEN_PREFIX, AND_PREFIX)
+
+
+def execute_one_step(context, name, **kwargs):
+    """
+    execute a step manually
+    :param name: step name
+    :param show: determine if the steps is displayed or not (the "Given" label is replaced to "And" label)
+    """
+    show = kwargs.get("show", False)
+    __logger__.debug("step defined in pre-actions: %s" % name)
+    if show:
+        print colored('    %s' % replace_given_to_add(name), 'green')
+    context.execute_steps(name)
 
 
 def before_all(context):
@@ -61,18 +118,20 @@ def before_feature(context, feature):
     :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     :param feature:
     """
+    global steps_after_feature, steps_before_scenario, steps_after_scenario
     __logger__.info("\n\n\n\n")
     __logger__.info(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>")
     __logger__.info(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>")
-    __logger__.info("BEFORE FEATURE: %s" % feature)
-    __logger__.info("BEFORE FEATURE: %s" % feature.description)
-    # ---- BackgroundFeature ----
-    keywords = ["Setup:", "Check:"]
-    for description in feature.description:
-        if description.split(" ")[0] in keywords:
-            description = description.replace(description.split(" ")[0], "Given")
-            __logger__.debug("steps in Background Feature: %s" % description)
-            context.execute_steps(description)
+    __logger__.info("FEATURE name: %s" % feature)
+    __logger__.info("FEATURE description: %s" % feature.description)
+    # ---- ConditionsBeforeFeature ----
+    steps_before_feature = get_steps_from_feature_description(ACTIONS_BEFORE_FEATURE, feature)
+    for item in steps_before_feature:
+        execute_one_step(context, item)
+
+    steps_after_feature = get_steps_from_feature_description(ACTIONS_AFTER_FEATURE, feature)
+    steps_before_scenario = get_steps_from_feature_description(ACTIONS_BEFORE_SCENARIO, feature)
+    steps_after_scenario = get_steps_from_feature_description(ACTIONS_AFTER_SCENARIO, feature)
 
 
 def after_feature(context, feature):
@@ -81,8 +140,10 @@ def after_feature(context, feature):
     :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     :param feature: feature properties
     """
-    context.execute_steps(u'Given stop service')
-    print colored('    And stop context Broker', 'green')
+    global steps_after_feature
+    # ---- ConditionsAfterFeature ----
+    for item in steps_after_feature:
+        execute_one_step(context, item, show=True)
     __logger__.info("AFTER FEATURE")
     __logger__.info("<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<")
     __logger__.info("<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<")
@@ -94,8 +155,12 @@ def before_scenario(context, scenario):
     :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     :param scenario: scenario properties
     """
+    global steps_before_scenario
     __logger__.info("==>>")
     __logger__.info("BEFORE SCENARIO: %s " % scenario)
+    # ---- Action Before each Scenario ----
+    for item in steps_before_scenario:
+        execute_one_step(context, item, show=True)
 
 
 def after_scenario(context, scenario):
@@ -105,8 +170,10 @@ def after_scenario(context, scenario):
     :param context: It’s a clever place where you and behave can store information to share around. It runs at three levels, automatically managed by behave.
     :param scenario: scenario properties
     """
-    context.execute_steps(u'Given delete database in mongo')
-    print colored('    And delete database in mongo', 'green')
+    global steps_after_scenario
+    # ---- Action After each Scenario ----
+    for item in steps_after_scenario:
+        execute_one_step(context, item, show=True)
     __logger__.info("AFTER SCENARIO")
     __logger__.info("<<==")
